+++ packages/datasource-mongoose/CHANGELOG.md
@@ -1,3 +1,13 @@
+## @forestadmin/datasource-mongoose [1.5.2-alpha-widgets.1](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-mongoose@1.5.1...@forestadmin/datasource-mongoose@1.5.2-alpha-widgets.1) (2023-08-23)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-toolkit:** upgraded to 1.10.0-alpha-widgets.1
+
 ## @forestadmin/datasource-mongoose [1.5.1](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-mongoose@1.5.0...@forestadmin/datasource-mongoose@1.5.1) (2023-08-11)
 
 
@@ -7,6 +17,26 @@
 ### Dependencies
 
 * **@forestadmin/datasource-toolkit:** upgraded to 1.9.0
+
+## @forestadmin/datasource-mongoose [1.5.1-alpha-widgets.2](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-mongoose@1.5.1-alpha-widgets.1...@forestadmin/datasource-mongoose@1.5.1-alpha-widgets.2) (2023-08-23)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-toolkit:** upgraded to 1.9.0-alpha-widgets.2
+
+## @forestadmin/datasource-mongoose [1.5.1-alpha-widgets.1](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-mongoose@1.5.0...@forestadmin/datasource-mongoose@1.5.1-alpha-widgets.1) (2023-08-23)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-toolkit:** upgraded to 1.9.0-alpha-widgets.1
 
 # @forestadmin/datasource-mongoose [1.5.0](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-mongoose@1.4.5...@forestadmin/datasource-mongoose@1.5.0) (2023-07-27)